requirements.txt
streamlit>=1.32.0

pandas>=2.0.0
numpy>=1.24.0

plotly>=5.18.0

firebase-admin>=6.4.0

google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

requests>=2.31.0

google-generativeai>=0.5.0

